Rename mock.js to mockUtility.js
